requirements.txt
flask>=3.0.0
requests>=2.31.0
python-dotenv>=1.0.0
mcp[cli]>=1.0.0
uvicorn>=0.27.0
starlette>=0.36.0